+++ src/SubRoutine.ts
@@ -1,6 +1,6 @@
 type subRoutineType = 'loop' | 'if';
 
-export class SubRoutineEnd {}
+export class SubRoutineEnd { }
 export class SubRoutine {
   type: subRoutineType;
   varName: string;
@@ -10,9 +10,11 @@ export class SubRoutine {
   functionName: string;
   loopIterator?: string;
   loopIndex?: string;
+  isTypescript: boolean;
 
-  constructor(type: subRoutineType, varName: string, liveId?: string) {
+  constructor(type: subRoutineType, varName: string, isTypescript: boolean, liveId?: string) {
     this.type = type;
+    this.isTypescript = isTypescript;
     this.varName = varName;
     this.children = [];
 
@@ -22,7 +24,16 @@ export class SubRoutine {
     } else if (liveId && liveId.charAt(0) === '#') {
       this.liveId = liveId.substring(1);
     }
-    this.functionName = `${type}_${this.varName.replace(/\W/g, '')}_${this.liveId || ''}`;
+    this.functionName = this.toCamelCase(`${type} ${this.varName.replace(/\W/g, ' ')} ${this.liveId || ''}`);
+  }
+
+  private toCamelCase(str: string) {
+    return str.replace(/(?:^\w|[A-Z]|\b\w|\s+)/g, function (match: string, index: number) {
+      if (+match === 0) {
+        return ""; // or if (/\s+/.test(match)) for white spaces
+      }
+      return index === 0 ? match.toLowerCase() : match.toUpperCase();
+    });
   }
 
   toString(): string {
@@ -33,25 +44,25 @@ export class SubRoutine {
         this.loopIterator = iterator;
         this.loopIndex = index;
         const liveLoopString = this.liveId
-          ? `self.register('${this.liveId}', { type: 'loop', node: elm, details: { startAt, fn, items, nodes } });\n`
+          ? `self.register('${this.liveId}', { type: 'loop', node: elm, details: { startAt, fn, fnName: '${this.functionName}', items, nodes } });\n`
           : '';
-
         return `{ 
-					const fn = ${this.functionName}.bind({},self, docElm, elm);
+          const fn = ${this.functionName}.bind({},self, docElm, elm);
 					const startAt = elm.childNodes.length;
-					const items = clone(self._getValue(self.data, '${varName}'));
+          const items = clone(self._getValue(self.data, '${varName}')) || [];
 					const nodes = fn(items);
 					${liveLoopString}
 				}`;
       case 'if':
         const liveIfString = this.liveId
-          ? `self.register('${this.liveId}', { type: 'conditional', node: elm, details: { startAt, fn, nodes, flag } });\n`
+          ? `self.register('${this.liveId}', { type: 'conditional', node: elm, details: { startAt, fn, fnName: '${this.functionName}', nodes, flag } });\n`
           : '';
         return ` {
 					const startAt = elm.childNodes.length;
 					const fn = ${this.functionName}.bind({},self, docElm, elm);
-					const flag = !!self._getValue(self.data, '${this.varName}')
-					const nodes = flag ? [fn()] : [];
+					const flag = !!self._getValue(self.data, '${this.varName}');
+					const nodes = flag ? fn() : [];
+
 					${liveIfString}
 				}`;
     }
@@ -60,16 +71,22 @@ export class SubRoutine {
   getFunction(): string {
     switch (this.type) {
       case 'loop':
-        return `function ${this.functionName} (self, docElm, elm, items) {
-	return self._forEach('${this.loopIterator}', '${this.loopIndex}',items, elm, function() { 
-		${this.children.join('\n')}
-	});
-}`;
+        const loopArgs = this.isTypescript ? 'self:JSNode, docElm:Document, elm:Node, items:any' : 'self, docElm, elm, items';
+        return `function ${this.functionName} (${loopArgs}) {
+          const fn = function() {
+            ${this.children.join('\n')}
+          };
+
+          return self._forEach('${this.loopIterator}', '${this.loopIndex}', elm, fn, items);
+        }`;
 
       case 'if':
-        return `function ${this.functionName} (self, docElm, elm) {
-	return getAddedChildren(elm, function () { ${this.children.join('\n')} })
-}`;
+        const ifArgs = this.isTypescript ? 'self:JSNode, docElm:Document, elm:Node' : 'self, docElm, elm';
+        return `function ${this.functionName} (${ifArgs}) {
+          const fn = function () { ${this.children.join('\n')} };
+	        return getAddedChildren(elm, fn);
+        }`;
+
     }
   }
 }