+++ CHANGELOG.md
@@ -42,3 +42,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - support for typescript (realized `.vscodeignore` prevents packing ts files)
 - getNode() now returns Node instead of the obscure JSNode
+
+## [1.0.3] - 2020-06-28
+
+### Added
+
+- Added more test, especially for typescript output
+
+### Fixed
+
+- improved support for loops (empty initial value + switching from arrays and objects)
+-